Back-end/MemoEthiopa/requirements.txt
Django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django-cors-headers
dj_database_url
Pillow
setuptools>=65.0
wheel>=0.38.4
drf-yasg
dotenv
pyotp
Procfile
waitress
gunicorn
whitenoise
drf-spectacular
langchain
langchain-groq
duckduckgo-search
langchain-community
groq
deep-translator
django-filter
